refactor: change usecase structure
