+++ src/domain/usecases/index.ts
@@ -1,3 +1,2 @@
 export * from "./getMovie.usecase";
 export * from "./getMovies.usecase";
-export * from "./getMovieSuggestion.usecase";
