+++ src/domain/usecases/interfaces/getMovieSuggestion.usecase.ts
@@ -1,0 +1,12 @@
+import { ListMoviesEntity, MovieEntity } from "../../entities/";
+
+export interface GetMovieSuggestionUsecase {
+  exec: (
+    params: GetMovieSuggestionUsecase.Params,
+  ) => Promise<GetMovieSuggestionUsecase.Model>;
+}
+
+export namespace GetMovieSuggestionUsecase {
+  export type Params = ListMoviesEntity;
+  export type Model = MovieEntity;
+}
