+++ src/domain/usecases/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from "./getMovie.usecase";
 export * from "./getMovies.usecase";
+export * from "./getMovieSuggestion.usecase";
