+++ src/main/factories/usecases/getMovie.factory.ts
@@ -1,6 +1,6 @@
-import { GetMovieUsecase } from "@/domain/usecases";
+import { GetMovieUsecase } from "@/domain/usecases/interfaces";
 import { makeMoviesRepository } from "../repositories/movies.factory";
-import { GetMovieUsecaseImpl } from "@/infrastructure/usecases";
+import { GetMovieUsecaseImpl } from "@/domain/usecases";
 
 export const makeGetMovieUsecase = (): GetMovieUsecase =>
   new GetMovieUsecaseImpl(makeMoviesRepository());
