+++ src/main/factories/usecases/getMovies.factory.ts
@@ -1,6 +1,6 @@
-import { GetMoviesUsecase } from "@/domain/usecases";
+import { GetMoviesUsecase } from "@/domain/usecases/interfaces";
 import { makeMoviesRepository } from "../repositories/movies.factory";
-import { GetMoviesUsecaseImpl } from "@/infrastructure/usecases";
+import { GetMoviesUsecaseImpl } from "@/domain/usecases";
 
 export const makeGetMoviesUsecase = (): GetMoviesUsecase =>
   new GetMoviesUsecaseImpl(makeMoviesRepository());
